+++ app/components/nitro_kit/combobox.rb
@@ -117,23 +117,7 @@ module NitroKit
           :input
         )
 
-        svg(
-          **slot_attributes(
-            :icon,
-            attributes: {
-              viewbox: "0 0 16 16",
-              width: 16,
-              height: 16,
-              fill: "none",
-              stroke: "currentColor",
-              stroke_width: 1.5,
-              stroke_linecap: "round",
-              stroke_linejoin: "round",
-              focusable: "false",
-              aria: { hidden: true }
-            }
-          )
-        ) { |svg| svg.path(d: "m4 6 4 4 4-4") }
+        render_in_slot(Icon.new(:chevron_down, size: :sm), :icon)
       end
     end