+++ app/components/nitro_kit/form_builder.rb
@@ -115,7 +115,7 @@ module NitroKit
       id: nil,
       label: I18n.t("nitro_kit.dropzone.label"),
       description: nil,
-      presentation: :input,
+      presentation: :minimal,
       direct_upload: true,
       multiple: false,
       accept: nil,